+++ apps/desktop/src/main/overlay-preload.js
@@ -1,0 +1,6 @@
+const { contextBridge, ipcRenderer } = require('electron');
+
+contextBridge.exposeInMainWorld('zenOverlay', {
+  close: () => ipcRenderer.send('zen:overlay:close'),
+  onShow: (callback) => ipcRenderer.on('zen:overlay:shown', () => callback())
+});